Split the ContaoFrameworkTest class accordingly.
